pimdbdata file splited . fetch has owns file
close #67
